Move arrayMove to utils
