+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/List/structureStrategies/FlatStructureStrategy.js
@@ -1,6 +1,6 @@
 // @flow
 import {action, computed, observable} from 'mobx';
-import {arrayMove} from '../../../components';
+import {arrayMove} from '../../../utils';
 import type {StructureStrategyInterface} from '../types';
 
 export default class FlatStructureStrategy implements StructureStrategyInterface {
