+++ src/Sulu/Bundle/AdminBundle/Resources/js/utils/index.js
@@ -1,8 +1,10 @@
 // @flow
+import {arrayMove} from 'react-sortable-hoc';
 import {buildQueryString} from './Request';
 import {translate} from './Translator';
 
 export {
+    arrayMove,
     buildQueryString,
     translate,
 };
