+++ src/Sulu/Bundle/PageBundle/Resources/js/containers/TeaserSelection/TeaserSelection.js
@@ -3,8 +3,9 @@ import React, {Fragment} from 'react';
 import {action, computed, observable} from 'mobx';
 import type {IObservableValue} from 'mobx';
 import {observer} from 'mobx-react';
-import {arrayMove, MultiItemSelection} from 'sulu-admin-bundle/components';
+import {MultiItemSelection} from 'sulu-admin-bundle/components';
 import {MultiListOverlay} from 'sulu-admin-bundle/containers';
+import {arrayMove} from 'sulu-admin-bundle/utils';
 import TeaserStore from './stores/TeaserStore';
 import Item from './Item';
 import teaserProviderRegistry from './registries/TeaserProviderRegistry';
